migrated ParterLogoLink files to .tsx and updated snaptshot
